Backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
Werkzeug==3.0.1
python-dotenv==1.0.0
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0
numpy==1.26.2
